BC-1422 - rename test page
